Fix kserve-router HTTP handlers configuration
There is a bug when setting up the kserve-router HTTP server endpoint handlers. Upstream project is using the default server mux, but when implementing auth we used a non-default handler.

When upstream introduced readiness probes, the code got mixed and the readiness probe endpoint is not taking effect. This commit fixes the configuration of the http server endpoints.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>