Fix kserve-controller creating two InferenceGraph pods
A finalizer is being added on the InferenceGraph CRD for proper clean-up. Since the finalizer code is after resolving URLs of the involved InferenceServices, the `path` operation over the IG resource would reset the in-memory IG struct object leading to a router-pod being created without the right command-line arguments. An immediate subsequent reconcile fixes the args, but a secondary pod is created.

Moving the configuration of the finalizer before resolving the ISVCs URLs fixes the double-pod creation.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>