feat: generic resource crud handling
Expanding on the idea of [create|update|delete]Object I changed it to
generic set of functions. This will prevent from accidental runtime
errors due to casting to the wrong type (see #688).

In addition to new function signatures our controller now embeds both
client.Client and record.EventRecorder to handle both interactions with
kube-apiserver as well as raising events on successful completion.

`Delete` operation ensures that the object we are about to delete is
actually owned by the owner object in a similar fashion as we do it when
updating.

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>