Path-based routing for model services (#712)
Since body-based routing is not yet available in llm-d and it's also
a more expensive operation, to direct a request to a specific model,
for now, we will match on a particular path prefix `/<namespace>/<name>`
and rewrite the URL for the backend with the empty `/` path to match
vLLM's expected path (`/v1/completions`, etc).

This is the default behavior and customers can add additional way of
reaching a specific model (body-based, host-based, etc).

We need to make sure that once we have the status update logic,
we handle match and filters configurations accordingly.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>