chore(test): splits existing/non-existing well known configs
There are some presets that are not defined yet, namely:

- `configPrefillWorkerPipelineParallelName`
- `configDecodeWorkerPipelineParallelName`
- `configWorkerPipelineParallelName`

To ensure that our tests capture the diff between existing files
(resources to be created in the cluster) and defined presets we rely on
when merging specs in the code the set has been split to reflect it.

The test has been adjusted to check if there is a diff between existing
files and defined constants in `config_merge.go` by keeping track of
tested files.

In addition, the test logic has been simplified by removing duplicate
loop over a map with expected preset content that verify actual
modifications based on the provided template.

Split from #723

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>