feat(ci): reworks prow test runner scripts (#801)
To facilitate llm-d specific installations, the scripts
has been reworked and aligned with upstream (gh-actions)
counterparts. Third parameter is treated as `DEPLOYMENT_PROFILE`
(defaulted to `serverless`) that will enable conditional
installation of llm-d specific pieces.

This change is backward compatible with new prow job that needs
to be adjusted as the first paramater is actually intended
to be pytest marker string that can be expression-based,
making it not feasible for profile detection.

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>
Signed-off-by: Jooho Lee <jlee@redhat.com>