feat(storage): add model storage support for multi-node LLM workloads (LWS)
Port the existing storage-related functionality from single-node workloads to support PVC, OCI, Hugging Face, S3 in multi-node setups.

Both leader and worker pods in the LeaderWorkerSet receive model artifact configuration, and also P/D configurations is covered.

The mutation of the `main` container command to specify the model location had to be removed from the code and hard-coded in the templates.  The motivation is because multi-node (LWS based) deployments are using a smarter start-up script which turns the command more complex. This effectively aligns LLMIsvc implementation to how InferenceServices work: the templates consistently use /mnt/models as the location for loading the model.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>