fix(webhook): relax HTTPRoute/Gateway validation
Users should be able to attach custom Gateway while still relying on
managed HTTPRoute. While controller already has the logic to support it,
the webhook was previously rejecting as invalid, expecting that "bring
your own" config is all-or-nothing scenario. This obviously is not
desired when users want simply to attach their LLMInferenceService to
another gateway.

- Webhook validation now permits such configuration, and only rejects
  when HTTPRoute Spec of LLMInferenceService is also set, treating it as
invalid configuration and suggesting removal of parentRef in favor of
the global Gateway defined as `.spec.router.gateway.refs` instead
- Adjusted EnvTests
- Introduced e2e test to cover this case that leverages llmd-simulator
  for faster feedback loop

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>